Remove redundant header file
